integration/genaiops-langfuse-on-aws/lab5/requirements.txt
bedrock-agentcore>=0.1.5
bedrock-agentcore-starter-toolkit>=0.1.13
langfuse>=3.5.1
python-dotenv>=1.1.1
strands-agents>=1.9.1
strands-agents-tools>=0.2.8
strands-agents[otel]>=1.9.1